Rename `TaskViewModule` -> `TaskDetailsModule`
